requirement.txt
Python >= 3.9
Pillow >= 9.0.0
